changed Names of Frames MakeReservation feature working Refactored some code
