+++ Database_Project/src/META-INF/MANIFEST.MF
@@ -1,3 +1,3 @@
 Manifest-Version: 1.0
-Main-Class: mainFrame
+Main-Class: launchAppFrame
